Make `UsageAnalyzer` a private class in `DdgGenerator`
